# docs/Jothidam_AI_OpenAPI_v1_Thirukanitham_2026.yaml
openapi: 3.0.3
info:
  title: Vinaadi AI API
  version: 1.0.0
  description: MVP API for Thirukanitham Tamil astrology companion with Family Aggregate Fortune in MVP 1.
servers:
  - url: https://api.jothidam.ai/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /panchangam/daily:
    get:
      summary: Get city-specific daily Panchangam
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: lat
          required: true
          schema: { type: number }
        - in: query
          name: lng
          required: true
          schema: { type: number }
        - in: query
          name: timezone
          required: true
          schema: { type: string }
      responses:
        '200': { description: Panchangam }
  /panchangam/timings:
    get:
      summary: Get Panchangam timing windows
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
        - in: query
          name: lat
          required: true
          schema: { type: number }
        - in: query
          name: lng
          required: true
          schema: { type: number }
        - in: query
          name: timezone
          required: true
          schema: { type: string }
      responses:
        '200': { description: Panchangam timing windows }
...
  /charts/{chartId}/gochar/current:
    get:
      summary: Get current transits from Moon and Lagna
      parameters:
        - in: path
          name: chartId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: datetime
          schema: { type: string, format: date-time }
        - in: query
          name: date
          schema: { type: string, format: date }
      responses:
        '200': { description: Current Gochar }
  /charts/{chartId}/sani-cycle:
    get:
      summary: Get current Saturn cycle status
      parameters:
        - in: path
          name: chartId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: date
          schema: { type: string, format: date }
      responses:
        '200': { description: Sani cycle }
...
  /family-vaults:
    get:
      summary: List family vaults for an owner
      parameters:
        - in: query
          name: ownerUserId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: limit
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        '200': { description: Family vault list }
    post:
      summary: Create family vault
      responses:
        '200': { description: Family vault created }
  /family-vaults/{familyVaultId}:
    get:
      summary: Get family vault metadata
      parameters:
        - in: path
          name: familyVaultId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Family vault metadata }
  /family-vaults/{familyVaultId}/members:
    post:
      summary: Add family member with birth profile
      parameters:
        - in: path
          name: familyVaultId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName, birthDateLocal, birthPlace, birthLatitude, birthLongitude, birthTimezone]
              properties:
                ownerUserId: { type: string, format: uuid }
                familyVaultId: { type: string, format: uuid, nullable: true }
                relationshipToOwner: { type: string, enum: [self, spouse, child, parent, sibling, grandparent, other] }
                displayName: { type: string }
                birthDateLocal: { type: string, format: date }
                birthTimeLocal: { type: string, example: '06:30:00' }
                birthPlace: { type: string }
                birthLatitude: { type: number }
                birthLongitude: { type: number }
                birthTimezone: { type: string, example: Asia/Kolkata }
                birthTimeConfidenceMinutes: { type: integer }
                calculateNow: { type: boolean, default: true }
                memberWeight: { type: number, default: 1.0 }
      responses:
        '200': { description: Family member created }
  /family-vaults/{familyVaultId}/daily-aggregate:
    get:
      summary: Get Family Aggregate Fortune dashboard for a date
      parameters:
        - in: path
          name: familyVaultId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: Family aggregate fortune }
  /family-vaults/{familyVaultId}/summary:
    get:
      summary: Get family summary card for a date
      parameters:
        - in: path
          name: familyVaultId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: Family summary card }
  /family-vaults/{familyVaultId}/calendar:
    get:
      summary: Get family aggregate fortune calendar
      parameters:
        - in: path
          name: familyVaultId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: from
          required: true
          schema: { type: string, format: date }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: Family aggregate calendar }
...
  /charts/{chartId}/peyarchi:
    get:
      summary: Get upcoming peyarchi events for chart
      parameters:
        - in: path
          name: chartId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Peyarchi events }
  /charts/{chartId}/peyarchi/upcoming:
    get:
      summary: Get peyarchi events within request window
      parameters:
        - in: path
          name: chartId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Upcoming peyarchi events }
...